requirements.txt
requests
python-dotenv
feedparser
tweepy
google-cloud-storage
pyyaml
